+++ inst/doc/Tips-tricks.R
@@ -7,6 +7,23 @@ knitr::opts_chunk$set(
 ## ----results='asis', message=FALSE--------------------------------------------
 library(surveytable)
 set_survey(namcs2019sv)
+tab_subset("AGER", "SPECCAT", test = "65-74 years")
+
+## ----echo=FALSE---------------------------------------------------------------
+version = packageVersion("surveytable")
+
+## ----results='asis', message=FALSE--------------------------------------------
+library(surveytable)
+set_survey(namcs2019sv)
+
+## ----echo=FALSE---------------------------------------------------------------
+set_opts(output = "raw")
+print( tab("SPECCAT", test = TRUE), destination = "")
+set_opts(reset = TRUE)
+
+## ----results='asis', message=FALSE--------------------------------------------
+library(surveytable)
+set_survey(namcs2019sv)
 
 ## ----results='asis'-----------------------------------------------------------
 tab_subset("NUMMED", "AGER")